requirements.txt
streamlit==1.32.2
pandas==2.2.1
pdfplumber==0.10.3
openpyxl==3.1.2
python-dotenv==1.0.0
Pillow==10.2.0 